requirements.txt
Flask==^2.0.3
python-binance==1.0.15